fix test but BlockFraudProofSubmitted test fails
