+++ contracts/rollup/IRollup.sol
@@ -20,6 +20,8 @@ interface IRollup {
 
 	error PairingCheckFailed();
 
+	error BlockNumberOutOfRange();
+
 	event DepositsProcessed(
 		uint256 indexed lastProcessedDepositId,
 		bytes32 depositTreeRoot
@@ -76,9 +78,9 @@ interface IRollup {
 		bytes32[] calldata depositHashes
 	) external;
 
-	function getBlockHashAndBuilder(
+	function getBlockBuilder(
 		uint256 blockNumber
-	) external view returns (bytes32, address);
+	) external view returns (address);
 
-	function getBlockHash(uint256 blockNumber) external view returns (bytes32);
+	function getBlockHash(uint32 blockNumber) external view returns (bytes32);
 }
